% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac.R
\name{ac}
\alias{ac}
\title{shortcut for as.character}
\usage{
ac(x)
}
\arguments{
\item{x}{object to turn into character}
}
\value{
as.character attempts to coerce its argument to character type
}
\description{
Change the class of an object to character
}
\examples{

as.character(5) #returns the number 5 as class 'character'
ac(5)           #returns the number 5 also as class 'character'

}
\references{
EU Lot 2 project
}
\seealso{
\code{\link{as.character}}
}
\author{
Niels T. Hintzen
}
