% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af.R
\name{af}
\alias{af}
\title{shortcut for as.factor}
\usage{
af(x)
}
\arguments{
\item{x}{object to turn into factor}
}
\value{
as.factor attempts to coerce its argument to factor type
}
\description{
Change the class of an object to factor
}
\examples{

as.factor(5)    #returns the number 5 as class 'factor'
af(5)           #returns the number 5 also as class 'factor'

}
\references{
EU Lot 2 project
}
\seealso{
\code{\link{as.factor}}
}
\author{
Niels T. Hintzen
}
