% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/an.R
\name{an}
\alias{an}
\title{shortcut for as.numeric}
\usage{
an(x)
}
\arguments{
\item{x}{object to turn into numeric}
}
\value{
as.numeric attempts to coerce its argument to numeric type
}
\description{
Change the class of an object to numeric
}
\examples{

as.numeric("5")   #returns the character 5 as class 'numeric'
an("5")           #returns the character 5 also as class 'numeric'

}
\references{
EU Lot 2 project
}
\seealso{
\code{\link{as.numeric}}
}
\author{
Niels T. Hintzen
}
