% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anf.r
\name{anf}
\alias{anf}
\title{shortcut for as.numeric(as.character())}
\usage{
anf(x)
}
\arguments{
\item{x}{object to turn from factor into numeric}
}
\value{
as.numeric attempts to coerce its argument to numeric type
}
\description{
Change the class of an object from factor to numeric
}
\examples{


res <- as.factor(5.1)
an(res)   #returns 1
anf(res)  #returns the original 5.1

}
\references{
EU Lot 2 project
}
\seealso{
\code{\link{as.numeric}}, \code{\link{as.character}}
}
\author{
Francois Bastardie
}
