% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bearing.R
\name{bearing}
\alias{bearing}
\title{Compute bearing between two points on a sphere (approximation of the earth)
at the starting point}
\usage{
bearing(lon, lat, lonRef, latRef)
}
\arguments{
\item{lon}{Longitude of point 1}

\item{lat}{Latitude of point 1}

\item{lonRef}{Longitude of point 2}

\item{latRef}{Latitude of point 2}
}
\description{
Compute the bearing between two GPS locations defined in longitude and
latitude notation on the earth. The earth is assumed to have a perfect
spherical shape. Bearing is returned in compass degrees.
}
\examples{

lon <- -4
lat <- 50
lonRef <- -4.2
latRef <- 51

bearing(lon,lat,lonRef,latRef) #352.8271

}
\references{
EU Lot 2 project, based on the Haversine formula, see also:
Hintzen et al. 2010 Fisheries Research
}
\seealso{
\code{\link{km2Degree}}, \code{\link{degree2Km}},
\code{\link{lonLatRatio}},\code{\link{distance}}
}
\author{
Niels T. Hintzen
}
