% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.listquote.r
\name{c.listquote}
\alias{c.listquote}
\title{Produce quoted list for use within data.table}
\usage{
\method{c}{listquote}(...)
}
\arguments{
\item{\dots}{character vector}
}
\value{
Returnes quoted list
}
\description{
Produce the quoted list format from a character vector which is used as
input for the data.table format
}
\details{
character vector could contain e.g. colnames(x) or c("COL1","COL2","COL3")
where COL1 etc are column names
}
\note{
With great thanks to original author.
}
\examples{

c.listquote(c("COL1","COL2","COL3")) #Returns: list(COL1,COL2,COL3)

}
\references{
See \code{\link{data.table}} documentation
}
\seealso{
See \code{\link{data.table}} documentation
}
\author{
Niels T. Hintzen
}
