% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tclWindows.r
\name{callPeakValue}
\alias{callPeakValue}
\title{Internal function to retreive information on number of peaks}
\usage{
callPeakValue(pks)
}
\arguments{
\item{pks}{Number of peaks spotted}
}
\description{
Internal function used together with analyseTacsatActivity.
}
\author{
Niels T. Hintzen
}
