\name{countPings}
\alias{countPings}
\title{Ping counting over trips}
\description{This function counts the number of pings in the tacsat data for each trip and then merges them with that same file}
\usage{countPings(tacsat = tacsat1f)}
\arguments{
  \item{data.frame, tacsat typically with only fishing pings identified}
}
\details{}
\value{}
\references{}
\author{Doug Beare}
\note{
}
\seealso{}
\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
