% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSquare.r
\name{CSquare}
\alias{CSquare}
\title{Calculate the CSquare notation from the GPS locations}
\usage{
CSquare(lon, lat, degrees)
}
\arguments{
\item{lon}{Longitudes of points}

\item{lat}{Latitudes of points}

\item{degrees}{Resolution of CSquare notation: 10, 5, 1, 0.5, 0.1, 0.05,
0.01}
}
\description{
Compute the CSquare notation from the GPS location where you can define what
resolution you want the CSquare notation to be.
}
\examples{

lon <- -4
lat <- 50
degrees <- 0.01

CSquare(lon,lat,degrees) # "7500:104:100:100"

}
\references{
EU Lot 2 project, based on CSquare:
http://www.marine.csiro.au/csquares/
}
\seealso{
\code{\link{km2Degree}}, \code{\link{degree2Km}},
\code{\link{lonLatRatio}}, \code{\link{distance}}
}
\author{
Niels T. Hintzen
}
