% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree2Km.R
\name{degree2Km}
\alias{degree2Km}
\title{Compute distance from degrees into kilometers}
\usage{
degree2Km(lon, lat, degree)
}
\arguments{
\item{lon}{Longitude of the GPS position}

\item{lat}{Latitude of the GPS positiona}

\item{degree}{Value in degrees to turn into Km}
}
\description{
Function transformes the distance expressed in degrees into kilometers. This
based on the GPS location of a point.
}
\note{
Computation of Km is approximation based on the Haversine formula
}
\examples{

lon <- -4
lat <- 50
degree <- 1.601833

degree2Km(lon,lat,degree) #114.4897km

}
\references{
EU lot 2 project
}
\seealso{
\code{\link{distance}}, \code{\link{km2Degree}},
\code{\link{lonLatRatio}}
}
\author{
Niels T. Hintzen
}
