% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Compute distance between two points on a sphere (approximation of the earth)}
\usage{
distance(lon, lat, lonRef, latRef)
}
\arguments{
\item{lon}{Longitude of point 2}

\item{lat}{Latitude of point 2}

\item{lonRef}{Longitude of point 1}

\item{latRef}{Latitude of point 1}
}
\description{
Compute the distance between two GPS locations defined in longitude and
latitude notation on the earth. The earth is assumed to have a perfect
spherical shape. Distance is returned in km.
}
\examples{

lon <- -4
lat <- 50
lonRef <- -4.2
latRef <- 51

distance(lon,lat,lonRef,latRef) #112.09

}
\references{
EU Lot 2 project, based on the Haversine formula, see also:
Hintzen et al. 2010 Fisheries Research
}
\seealso{
\code{\link{km2Degree}}, \code{\link{degree2Km}},
\code{\link{lonLatRatio}}
}
\author{
Niels T. Hintzen
}
