% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceTacsat.r
\name{distanceTacsat}
\alias{distanceTacsat}
\title{Calculate the distance (in km) from a Tacsat dataset}
\usage{
distanceTacsat(tacsat, index)
}
\arguments{
\item{tacsat}{tacsat (normal or high ping rate) dataset}

\item{index}{Matrix with specification of succeeding datapoints (see details
for format)}
}
\description{
Calculates the distance of a Tacsat dataset with specification of succeeding
datapoints. Distance is only calculated between these specified datapoints
}
\details{
index is designed as a matrix where rows represent succeeding datapoints,
column 1 represent start points, column 2 represent end points.
}
\examples{


data(tacsat)
#Speed threshold points (two values), NULL means use all points
st        <- c(2,6)
#Remove duplicate records in VMS dataset
remDup    <- TRUE

  #Sort the VMS data
tacsat     <- sortTacsat(tacsat)
tacsat     <- tacsat[1:1000,]

  #Filter the VMS data
tacsat     <- filterTacsat(tacsat,st,NULL,remDup)

distanceTacsat(tacsat,matrix(c(2,3,3,4),nrow=2,ncol=2,
                dimnames=list(1:2,c("startpoint","endpoint"))))
#6.335944 14.847291

}
\references{
EU lot 2 project
}
\seealso{
\code{\link{distance}}, \code{\link{distanceInterpolation}},
\code{\link{diffInter}}
}
\author{
Niels T. Hintzen
}
