% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmstools-package.R
\docType{data}
\name{eflalo}
\alias{eflalo}
\title{Vessel logbook dataset}
\format{
A data frame with 4539 observations on the following 189 variables.
\describe{ \item{list("VE_REF")}{Vessel reference / ID}
\item{list("VE_FLT")}{Fleet reference (DCF regulation)}
\item{list("VE_COU")}{Flag nation of vessel} \item{list("VE_LEN")}{Vessel
length} \item{list("VE_KW")}{Vessel power (in kW)}
\item{list("VE_TON")}{Vessel Tonnage (GT)} \item{list("FT_REF")}{Fishing
trip reference number} \item{list("FT_DCOU")}{Departure country}
\item{list("FT_DHAR")}{Departure harbour} \item{list("FT_DDAT")}{Departure
date} \item{list("FT_DTIME")}{Departure time} \item{list("FT_LCOU")}{Landing
country} \item{list("FT_LHAR")}{Landing harbour}
\item{list("FT_LDAT")}{Landing date} \item{list("FT_LTIME")}{Landing time}
\item{list("LE_ID")}{Log event ID} \item{list("LE_CDAT")}{Catch date}
\item{list("LE_STIME")}{Log event start time} \item{list("LE_ETIME")}{Log
event end time} \item{list("LE_SLAT")}{Log event start position latitude}
\item{list("LE_SLON")}{Log event start position longitude}
\item{list("LE_ELAT")}{Log event end position latitude}
\item{list("LE_ELON")}{Log event end position longitude}
\item{list("LE_GEAR")}{Gear} \item{list("LE_MSZ")}{Mesh size}
\item{list("LE_RECT")}{ICES rectangle} \item{list("LE_DIV")}{ICES division}
\item{list("LE_MET_level6")}{Fishing activity}
\item{list("LE_KG_SPEC")}{Landing weight estimate of species (FAO species
codes)} \item{list("LE_EURO_SPEC")}{Landing value of species (FAO species
codes)} }
}
\source{
NA
}
\description{
A dataset consisting of logbook (landings / values) registrations of
disguised origin for 2 consecutice years.
}
\details{
This example dataset if for example purposes only. Without prior permission
of the authors it is not allowed to use this data other than for example
non-publishable purposes.
}
\examples{

data(eflalo)
eflalo <- formatEflalo(eflalo)
str(eflalo)

}
\references{
EU Lot 2 project, Niels T. Hintzen (niels.hintzen@wur.nl)
}
\keyword{datasets}
