% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eflalo2relational.r
\name{eflalo2relational}
\alias{eflalo2relational}
\title{Turn eflalo dataset into relational format}
\usage{
eflalo2relational(x)
}
\arguments{
\item{x}{Dataframe with eflalo data and eflalo format}
}
\description{
Turn the column setup of eflalo into a row setup where each species catch
has its own row
}
\details{
May take a long time for long eflalo datasets
}
\examples{

data(eflalo)
eflalo    <- eflalo[1:20,]
eflaloRel <- eflalo2relational(eflalo)

}
\references{
EU Lot 2 project
}
\seealso{
\code{\link{formatEflalo}},\code{\link{readEflalo}}
}
\author{
Niels T. Hintzen
}
