% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eflaloHaul2Tacsat.r
\name{eflaloHaul2Tacsat}
\alias{eflaloHaul2Tacsat}
\title{Convert eflalo (haul by haul / log-event) to tacsat format}
\usage{
eflaloHaul2Tacsat(x, npoints = NULL)
}
\arguments{
\item{x}{eflalo formatted dataset with haul-by-haul data (e.g.
LE_STIME,LE_SLON)}

\item{npoints}{Number of intermediate points to include (defaults to 10)}
}
\value{
Tacsat formatted dataset is returned
}
\description{
Convert haul by haul / log event organised eflalo data to tacsat relational
style
}
\seealso{
\code{\link{eflalo2Relational}}
}
\author{
Niels T. Hintzen
}
