% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmstools-package.R
\docType{data}
\name{europa}
\alias{europa}
\title{Geometry dataset of Europe}
\format{
\describe{ \item{geometry}{
}
\source{
NOAA coastline datafile
}
\description{
Geometry dataset with the coastlines of Europe.
}
\examples{

data(europa)
plot(st_geometry(eurPols),col="green",xlim=c(-4,10),ylim=c(48,62))

}
\keyword{datasets}
