% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatEflalo.r
\name{formatEflalo}
\alias{formatEflalo}
\title{Format Eflalo data to ensure that all columns are in right format}
\usage{
formatEflalo(x)
}
\arguments{
\item{x}{eflalo file}
}
\value{
Returns original Eflalo file but now with reformatted data
}
\description{
Reformat all the columns of the Eflalo data to ensure that all data is in
the right format
}
\examples{

data(eflalo)
eflalo <- formatEflalo(eflalo)

}
\references{
EU lot 2 project
}
\seealso{
\code{\link{formatTacsat}}
}
\author{
Niels T. Hintzen
}
