% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatTacsat.r
\name{formatTacsat}
\alias{formatTacsat}
\title{Format tacsat data to ensure that all columns are in right format}
\usage{
formatTacsat(x)
}
\arguments{
\item{x}{tacsat file}
}
\value{
Returns original tacsat file but now with reformatted data
}
\description{
Reformat all the columns of the tacsat data to ensure that all data is in
the right format
}
\examples{

data(tacsat)
tacsat <- formatTacsat(tacsat)

}
\references{
EU lot 2 project
}
\seealso{
\code{\link{formatEflalo}}
}
\author{
Niels T. Hintzen
}
