% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICESarea.r
\name{ICESarea}
\alias{ICESarea}
\title{Get ICES area from coordinates}
\usage{
ICESarea(tacsat, areas, st_crs = NULL, fast = FALSE)
}
\arguments{
\item{tacsat}{dataframe given that they have 'SI_LONG' and 'SI_LATI' columns
(either tacsat format or other dataset with SI_LONG and SI_LATI columns)}

\item{areas}{ICES areas as SpatialPolygons}

\item{fast}{If memory allocation is not a problem, a faster version can be
switched on}

\item{proj4string}{Projection string, default to NULL.}
}
\value{
Returns the areas as a vector
}
\description{
Get the ICES area from any lon,lat position, given that this position is
within the ICES region.
}
\examples{

data(ICESareas)
res   <- data.frame(SI_LONG = c(1,2,2,4,2),
                    SI_LATI = c(53,53.2,54,56.7,55.2))
areas <- ICESarea(res,ICESareas,st_crs=4326)
ICESareas$Area_Full[areas]

}
\references{
EU Lot 2 project
}
\seealso{
\code{\link{ICESrectangle}}, \code{\link{ICESrectangle2LonLat}}
}
\author{
Niels T. Hintzen
}
