% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICESrectangle2LonLat.r
\name{ICESrectangle2LonLat}
\alias{ICESrectangle2LonLat}
\title{Convert the ICES rectangle to coordinates}
\usage{
ICESrectangle2LonLat(statsq, midpoint = FALSE)
}
\arguments{
\item{statsq}{character vector of ICES rectangles}

\item{midpoint}{logical statement if returned values need to be midpoints of
rectangle or not (default)}
}
\value{
Returns dataframe with longitude and latitude columns
}
\description{
Convert the ICES rectangle to the longitude and latitude position of that
specific rectangle, with either midpoint or lower-left corner.
}
\examples{

res       <- data.frame(SI_LONG = c(1,2,2,4,2),
                        SI_LATI = c(53,53.2,54,56.7,55.2))
rects     <- ICESrectangle(res)
midpoints <- ICESrectangle2LonLat(rects,midpoint=TRUE)
corners   <- ICESrectangle2LonLat(rects,midpoint=FALSE)


}
\references{
EU Lot 2 project
}
\seealso{
\code{\link{ICESarea}}, \code{\link{ICESrectangle}}
}
\author{
Neil Campbell, Mark Payne
}
