% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interStraightLine.r
\name{interStraightLine}
\alias{interStraightLine}
\title{Interpolate using Straight line}
\usage{
interStraightLine(spltx, spltCon, res)
}
\arguments{
\item{spltx}{Tacsat dataset split up by vessel using 'split'.}

\item{spltCon}{List of VMS records that are connected}

\item{res}{Number of additional points to use for interpolation}
}
\description{
This is an internal function of interpolateTacsat. Function to interpolate
VMS data using a straight line.
}
\author{
Niels T. Hintzen
}
