% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km2Degree.R
\name{km2Degree}
\alias{km2Degree}
\title{Compute distance from kilometers into degrees}
\usage{
km2Degree(lon, lat, km)
}
\arguments{
\item{lon}{Longitude of the GPS position}

\item{lat}{Latitude of the GPS positiona}

\item{km}{Value in Km to turn into degrees}
}
\description{
Function transformes the distance expressed in kilometers into degrees. This
based on the GPS location of a point.
}
\note{
Computation of degrees is approximation based on the Haversine formula
}
\examples{

lon <- -4
lat <- 50
km  <- 114.4897

km2Degree(lon,lat,km) #1.601833

}
\references{
EU lot 2 project
}
\seealso{
\code{\link{distance}}, \code{\link{degree2Km}},
\code{\link{lonLatRatio}}
}
\author{
Niels T. Hintzen
}
