% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lonLat2SFPolygons.r
\name{lonLat2SFPolygons}
\alias{lonLat2SFPolygons}
\title{Creates a 'SF polygon' object from longitude-latitude dataset}
\usage{
lonLat2SFPolygons(SI_LONG = NULL, SI_LATI = NULL, lst = NULL)
}
\arguments{
\item{SI_LONG}{Set with longitude values}

\item{SI_LATI}{Set with latitude values}

\item{lst}{list of objects, each with longitude and latitude values}
}
\value{
Returns a 'sfc' object from the package 'sf'
}
\description{
Takes one set of longitude-lattitude or a list of lon-lat objects and
converts it into one sfc object.
}
\examples{

data(europa)

eurPols     <- lonLat2SpatialPolygons(lst=lapply(as.list(sort(unique(europa$SID))),
                      function(x){data.frame(
                         SI_LONG=subset(europa,SID==x)$X,
                         SI_LATI=subset(europa,SID==x)$Y)}))

area        <- lonLat2SpatialPolygons(SI_LONG=c(2,2.5,2.7,2.1),SI_LATI=c(54,54.2,55.8,55.6))
plot(eurPols,col="green",xlim=c(-4,10),ylim=c(48,62)); axis(1);axis(2);box()
plot(area,add=TRUE,col="red")


}
\references{
EU Lot 2 project
}
\seealso{
\code{\link{st_polygon}}, \code{\link{st_multipolygon}}
}
\author{
Niels T. Hintzen
}
