% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lonLatRatio.r
\name{lonLatRatio}
\alias{lonLatRatio}
\title{Calculate the ratio between 1 degree in longitude versus 1 degree in
latitude}
\usage{
lonLatRatio(x1, lat)
}
\arguments{
\item{lat}{Latitude of the two GPS positions}

\item{lon}{Longitude of the two GPS positions}
}
\description{
The distance in Km on the longitude direction changes along the latitude
direction. This function computes the ratio between 1 degree in the
longitude direction depending on the latitude of the GPS position. Returns
the ratio's of two GPS locations (two succeeding VMS datapoints). Can be
used with 1 GPS position too, return NA for second value.
}
\note{
Computation is approximation based on the Haversine formula
}
\examples{

lon <- -4
lat <- 50

lonLatRatio(lon,lat)

}
\references{
EU lot 2 project
}
\seealso{
\code{\link{distance}}, \code{\link{degree2Km}},
\code{\link{km2Degree}}
}
\author{
Niels T. Hintzen
}
