\name{mcp.area}
\alias{mcp.area}
\title{
Hidden function to calculate Minimum Convex Polygon}
\description{
}
\usage{
mcp.area(xy, id, percent = seq(20, 100, by = 5), unin = c("m", "km"), unout = c("ha", "km2", "m2"), plotit = TRUE)
}
\arguments{
  \item{xy}{
}
  \item{id}{
}
  \item{percent}{
}
  \item{unin}{
}
  \item{unout}{
}
  \item{plotit}{
}
}
\author{
adehabitat
}
