\name{poolEflaloSpecies}
\alias{poolEflaloSpecies}
\title{ merge species in eflalo to reduce the size of data according to a threshold in kg}
\description{
merge species in eflalo to reduce the size of data according to a threshold in kg
}
\usage{
poolEflaloSpecies(eflalo, threshold.value, code)
}

\arguments{
  \item{eflalo}{data.frame, eflalo format
}
  \item{threshold.value}{numeric, in euro
}
  \item{code}{character, FAO code for the species other
}
}
\author{
Francois Bastardie
}

