% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbindEflalo.r
\name{rbindEflalo}
\alias{rbindEflalo}
\title{Row bind two eflalo sets together}
\usage{
rbindEflalo(set1, set2)
}
\arguments{
\item{set1}{Eflalo set 1}

\item{set2}{Eflalo set 2}
}
\description{
Row binds two eflalo sets together while taking differences in column names
into account
}
\examples{

data(eflalo)
set1 <- eflalo
set2 <- eflalo[seq(1,100,5),]

combined <- rbindEflalo(set1,set2)


}
\references{
EU Lot 2 project
}
\seealso{
\code{\link{rbindTacsat}}, \code{\link{do.call}}
}
\author{
Niels T. Hintzen
}
