% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortTacsat.r
\name{sortTacsat}
\alias{sortTacsat}
\title{Sorting Tacsat dataframe}
\usage{
sortTacsat(dat)
}
\arguments{
\item{dat}{tacsat dataframe}
}
\description{
Sort the Tacsat data first by vessel, then by date, speed and heading. Needs
to be in this order to be effectively used in other EU lot 2 project generic
functions.
}
\note{
Uses library(doBy)
}
\examples{

data(tacsat)
require(doBy)

  #Sort the Tacsat data
tacsat     <- sortTacsat(tacsat)

}
\references{
EU lot 2 project
}
\seealso{
\code{\link{filterTacsat}}
}
\author{
Niels T. Hintzen
}
