% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeTacsat.r
\name{summarizeEflalo}
\alias{summarizeEflalo}
\title{Summarize primary aspects of eflalo file}
\usage{
summarizeEflalo(eflalo)
}
\arguments{
\item{eflalo}{eflalo dataframe}
}
\description{
Summary nr vessels, countries, temporal range, effort, top 3 catches and
values
}
\examples{

data(eflalo)

summarizeEflalo(eflalo)

}
\seealso{
\code{\link{summarizeTacsat}},\code{\link{summary}}
}
\author{
Niels T. Hintzen
}
