% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeTacsat.r
\name{summarizeTacsat}
\alias{summarizeTacsat}
\title{Summarize primary aspects of tacsat file}
\usage{
summarizeTacsat(tacsat)
}
\arguments{
\item{tacsat}{tacsat dataframe}
}
\description{
Summary nr vessels, countries, spatial range, temporal range, effort from
tacsat
}
\examples{

data(tacsat)

summarizeTacsat(tacsat)

}
\seealso{
\code{\link{summarizeEflalo}},\code{\link{summary}}
}
\author{
Niels T. Hintzen
}
