% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmstools-package.R
\docType{data}
\name{tacsat}
\alias{tacsat}
\title{Vessel Monitoring by Satellite system dataset}
\format{
A data frame with 97015 observations on the following 8 variables.
\describe{ \item{list("VE_COU")}{Flag nation of vessel}
\item{list("VE_REF")}{Vessel reference / ID}
\item{list("SI_LATI")}{Latitudinal position (in decimal degrees)}
\item{list("SI_LONG")}{Longitudinal position (in decimal degrees)}
\item{list("SI_DATE")}{Date} \item{list("SI_TIME")}{Time (24 hours clock)}
\item{list("SI_SP")}{Speed (in knots per hour)} \item{list("SI_HE")}{Heading
(in degrees)} }
}
\source{
NA
}
\description{
A dataset consisting of VMS registrations of disguised origin for 2
consecutice years.
}
\details{
This example dataset if for example purposes only. Without prior permission
of the authors it is not allowed to use this data other than for example
non-publishable purposes.
}
\examples{

data(tacsat)
tacsat <- formatTacsat(tacsat)
str(tacsat)

}
\references{
EU Lot 2 project, Niels T. Hintzen (niels.hintzen@wur.nl)
}
\keyword{datasets}
