% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmstools-package.R
\docType{data}
\name{VMShf}
\alias{VMShf}
\title{Small high frequency test VMS dataset CANNOT BE DISTRIBUTED WITHOUT PRIOR
PERMISSION -> SEE Author}
\format{
A data frame with 702 observations on the following 6 variables.
\describe{ \item{list("ship")}{ID of ship} \item{list("declat")}{Decimal
notation of longitude position} \item{list("declon")}{Decimal notation of
latitude postion} \item{list("speed")}{Speed at ping}
\item{list("heading")}{Heading at ping} \item{list("date")}{Date and time of
ping} }
}
\source{
niels.hintzen@wur.nl
}
\description{
A short VMS dataset with high frequency interval rate (every 6 minutes) used
to test spatial and VMS functions. Dataframe contains vessel ID, decimal
longitude and decimal latitude position, speed and heading at ping and date
of ping.
}
\examples{

data(VMShf)
str(VMShf)


}
\references{
niels.hintzen@wur.nl
}
