\name{VMStoolsIMARES-package}
\alias{VMStoolsIMARES-package}
\alias{VMStoolsIMARES}
\docType{package}
\title{
development of R scripts to setup the workflow of vms and logbook data to compute pressure indicators
}
\description{
-Design of routines to import VMS and Logbook data\\*
-Design of routines to cluster metiers\\*
-Design of routines to link VMS and Logbook data\\*
-Design of routines to define vessel activity \\*
-Design of routines to interpolate VMS data \\*
-Design of routines to compute pressure indicators
}
\details{
\tabular{ll}{
Package: \tab VMStoolsIMARES\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-05-13\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Niels T. Hintzen, Doug Beare, Francois Bastardie, Nicolas Deporte, Andy South, Neil Campbell

Maintainer: Niels T. Hintzen <niels.hintzen@wur.nl>
}
\references{
}
