\name{callPeakValue}
\alias{callPeakValue}
\title{
Internal function to retreive information on number of peaks}
\description{
Internal function used together with analyseTacsatActivity.
}
\usage{
callPeakValue(pks)
}
\arguments{
  \item{pks}{
Number of peaks spotted}
}
\author{
Niels T. Hintzen}
