% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmstools-package.R
\docType{data}
\name{europa}
\alias{europa}
\title{Polygon dataset of Europe}
\format{
A data frame with 83878 observations on the following 5 variables.
\describe{ \item{list("PID")}{a numeric vector} \item{list("SID")}{a numeric
vector} \item{list("POS")}{a numeric vector} \item{list("X")}{Longitude
values} \item{list("Y")}{Latitude values} }
}
\source{
NOAA coastline datafile
}
\description{
Polygon dataset with the coastlines of Europe.
}
\examples{

data(europa)
require(PBSmapping)

eurPols     <- lonLat2SpatialPolygons(lst=lapply(as.list(sort(unique(europa$SID))),
                        function(x){data.frame(
                          SI_LONG=subset(europa,SID==x)$X,
                          SI_LATI=subset(europa,SID==x)$Y)}))

plot(eurPols,col="green",xlim=c(-4,10),ylim=c(48,62))

}
\keyword{datasets}
