% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmstools-package.R
\docType{data}
\name{ICESareas}
\alias{ICESareas}
\title{SpatialPolygons dataset of ICESareas}
\source{
www.ices.dk
}
\description{
A pre-defined SpatialPolygons dataset with the ICES areas.
}
\examples{

require(sp)
data(ICESareas)
plot(ICESareas)

}
\keyword{datasets}
