\name{ICESareas}
\alias{ICESareas}
\docType{data}
\title{
SpatialPolygons dataset of ICESareas}
\description{
A pre-defined SpatialPolygons dataset with the ICES areas.
}
\usage{data(ICESareas)}
\source{
www.ices.dk}
\examples{
require(sp)
data(ICESareas)
plot(ICESareas)
}
\keyword{datasets}
