% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgeur.r
\name{kgeur}
\alias{kgeur}
\title{Return kg and euro column index of eflalo dataset}
\usage{
kgeur(x)
}
\arguments{
\item{x}{Colnames of eflalo dataset (or any other dataset with column names
as LE_KG_ and LE_EURO_)}
}
\description{
Returns the index of the columns with kg and euro information from a given
eflalo dataset
}
\examples{

data(eflalo)
kgeur(colnames(eflalo))

}
\references{
EU Lot 2 project
}
\author{
Niels T. Hintzen
}
