% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointOnLand.r
\name{pointOnLand}
\alias{pointOnLand}
\title{Find points on land given a set of coordinates}
\usage{
pointOnLand(tacsat, lands, st_crs = NULL)
}
\arguments{
\item{tacsat}{Tacsat file}

\item{lands}{Polygon of area that is considered to be land}

\item{st_crs}{Projection string, default to NULL.}
}
\value{
Returns a vector with values 0 and 1. 1 indicating points on land, 0
indicating points not on land.
}
\description{
Find the points that are on land given a set of coordinates and polygons
that determine the part that is land
}
\details{
With many coordinates, the checking might take longer.
}
\examples{


data(tacsat)
data(europa)
tacsat  <- tacsat[1:1000,]
tacsat  <- sortTacsat(tacsat)

idx     <- pointOnLand(tacsat,europa,st_crs=4326);
idx     <- which(idx == 1)

plot(st_geometry(europa),xlim=c(0,10),ylim=c(48,62))
points(tacsat$SI_LONG[idx],tacsat$SI_LATI[idx],col="red",cex=0.5,pch=19)

}
\references{
EU Lot 2 project
}
\seealso{
\code{\link{pointInHarbour}}
}
\author{
Niels T. Hintzen
}
