\name{VMStools-package}
\alias{VMStools-package}
\alias{VMStools}
\docType{package}
\title{
development of R scripts to setup the workflow of vms and logbook data
}
\description{
-Design of routines to import VMS and Logbook data\cr
-Design of routines to cluster metiers\cr
-Design of routines to link VMS and Logbook data\cr
-Design of routines to define vessel activity\cr
-Design of routines to interpolate VMS data\cr
-Design of routines to compute pressure indicators\cr
}
\details{
\tabular{ll}{
Package: \tab VMStools\cr
Type: \tab Package\cr
Version: \tab 0.58\cr
Date: \tab 2011-12-30\cr
License: \tab \cr
LazyLoad: \tab yes\cr
}}
\author{
Niels T. Hintzen, Doug Beare, Francois Bastardie, Nicolas Deporte, Andy South, Neil Campbell, Fabrizio Manco, Clara Ulrich Rescan,
Stephanie Mahevas, Sebastien Demaneche, Josefine Egekvist, Hans Gerritsen, Matthew Parker-Humphreys

Maintainer: Niels T. Hintzen <niels.hintzen@wur.nl>, Francois Basterdie <fba@aqua.dtu.dk>
}
\references{EU Lot 2 project
}
