\name{VMStoolsIMARES-package}
\alias{VMStoolsIMARES-package}
\alias{VMStoolsIMARES}
\docType{package}
\title{
A set of tools to analyse raw Vessel Monitoring System (VMS) data, to couple VMS positions to logbooks,
to map the fishing effort and landings on fine scale, to format data toward the fishframe data warehouse.  
}
\description{
Logbook analyses to allocate the fishing activity due to various fisheries are integrated with processing of
raw satellite-recorded data for identifying trips at sea and fishing sequences.
Both data sources are linked into one output dataset.
A method is proposed to allocate logbooks landings to VMS positions. Some mapping facilities are also provided. 
The merged outcome is designed to direclty inform the FishFrame 
data warehouse encapuslating additional fishery data and reporting tools.
}
\details{
\tabular{ll}{
Package: \tab VMStoolsIMARES\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-05-13\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}

}
\author{
Niels Hintzen <niels.hintzen@wur.nl>, Doug Beare <doug.beare@wur.nl>, Francois Bastardie <fba@aqua.dtu.dk>

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
Bastardie, F., Nielsen, J. R., Ulrich, C., Egekvist, J. and Degel, H. 2010. 
Detailed mapping of fishing effort and landings by coupling fishing logbooks with satellite-recorded vessel geo-location. 
Fisheries Research,  106: 41-53.

Hintzen, N.T., Piet, G.J., Brunel, T., 2010. 
Improved estimation of trawling tracks using cubic Hermite spline interpolation of position registration data.
Fish. Res. 101, 108-115.
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
